requirements/framework.txt
accelerate
addict
aiohttp
attrdict
binpacking
charset_normalizer
cpm_kernels
dacite
datasets>=3.0,<4.0
einops
fastapi
gradio>=3.40.0,<6.0
importlib_metadata
json_repair
matplotlib
modelscope>=1.23
nltk
numpy
openai
oss2
pandas
peft>=0.11,<0.19
pillow
PyYAML>=5.4
requests
rouge
safetensors
scipy
sentencepiece
simplejson>=3.3.0
sortedcontainers>=1.5.9
tensorboard
tiktoken
tqdm
transformers>=4.57.0,<5.6.0
transformers_stream_generator
trl>=0.15,<0.30
uvicorn
zstandard
